all indexes and sizes to int

+++ lib/mtxLib.h
@@ -38,7 +38,7 @@
 #define NROWS(arr) sizeof(arr) / sizeof(arr[0])
 #define COLXROW(arr) sizeof(arr) / sizeof(arr[0][0])
 
-enum mtxResultInfo mtx_chol1(double *A, double *L, const ptrdiff_t size);
+enum mtxResultInfo mtx_chol1(double *A, double *L, const int size);
 
 enum mtxResultInfo
 {
@@ -52,22 +52,22 @@ enum mtxResultInfo
 
 typedef struct sMatrix
 {
-	ptrdiff_t nelem;
-	ptrdiff_t nrow;
-	ptrdiff_t ncol;
+	int nelem;
+	int nrow;
+	int ncol;
 	double *val;
 } Matrix_t;
 
 typedef struct sMatrixBool
 {
-	ptrdiff_t nelem;
-	ptrdiff_t nrow;
-	ptrdiff_t ncol;
+	int nelem;
+	int nrow;
+	int ncol;
 	_Bool *val;
 } MatrixBool_t;
 
-extern void mtx_init_bool(MatrixBool_t *const pSrc, _Bool *const pValue, const ptrdiff_t nrow, const ptrdiff_t ncol, const ptrdiff_t nelem);
-extern void mtx_init(Matrix_t *const pSrc, double *const pValue, const ptrdiff_t nrow, const ptrdiff_t ncol, const ptrdiff_t nelem);
+extern void mtx_init_bool(MatrixBool_t *const pSrc, _Bool *const pValue, const int nrow, const int ncol, const int nelem);
+extern void mtx_init(Matrix_t *const pSrc, double *const pValue, const int nrow, const int ncol, const int nelem);
 extern enum mtxResultInfo mtx_mul(Matrix_t const *const pSrc1, Matrix_t const *const pSrc2, Matrix_t *const pDst);
 extern enum mtxResultInfo mtx_transp_square(Matrix_t *const pSrc);
 extern enum mtxResultInfo mtx_transp_dest(Matrix_t const *const pSrc, Matrix_t *const pDst);
